TST: Moved files for fracture mesh generation
